requirements.txt
python-docx>=1.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
pdfplumber>=0.11.0
anthropic>=0.40.0
python-multipart>=0.0.9
openai>=1.30.0
google-generativeai>=0.8.0
python-dotenv>=1.0.1
openpyxl>=3.1.0
